+++ CHANGELOG.md
@@ -5,16 +5,21 @@
 * New event `immers-client-profile-update` fires whenever logged-in user's profile
 has changed
 * New method `immersClient.waitUntilConnected`, utility to simplify timing checks when a logged-in user is required (contributed by @wswoodruff)
-* Added `avatarId` prop to `Profile`
+* Added `bio` and `avatarObject` to `Profile`
+* Added `inbox` and `outbox` to `Profile.collections`
+* `ImmersClient.DestinationFromPlace` static method to transform ActivityPub Place to Destination type
 
 ### Changed
 
 * `URLFromProperty` can also find `href` prop if given a `Link` object
+* `Activities.getObject` can accept URL object in addition to string
+* Destination urls ending with an empty fragment will have the trailing `#` dropped to more accutately aggregate Destinations history
 
 ### Fixed
 
 * Restore summary messages in arrive/leave activities
 * Fix local immer place object not updating
+* Fix Profile type, url for avatar 3d Model is in prop `avatarModel` not `avatarGltf`
 
 ## v2.7.1 (2022-08-04)